optimizations for jdtls.lua
